+++ noisemodelling-scripts/src/test/groovy/org/noise_planet/noisemodelling/scripts/TestAcousticTools.groovy
@@ -20,7 +20,6 @@ import org.h2gis.utilities.TableLocation
 import org.noise_planet.noisemodelling.jdbc.NoiseMapDatabaseParameters
 import org.noise_planet.noisemodelling.scripts.Acoustic_Tools.Add_Laeq_Leq_columns
 import org.noise_planet.noisemodelling.scripts.Acoustic_Tools.Create_Isosurface
-import org.noise_planet.noisemodelling.scripts.NoiseModelling.Noise_level_from_traffic
 import org.noise_planet.noisemodelling.scripts.NoiseModelling.Road_Emission_from_Traffic
 import org.noise_planet.noisemodelling.scripts.Receivers.Delaunay_Grid
 import org.slf4j.Logger
