+++ noisemodelling-scripts/src/test/groovy/org/noise_planet/noisemodelling/scripts/TestNoiseModelling.groovy
@@ -114,78 +114,6 @@ class TestNoiseModelling extends JdbcTestCase {
     }
 
     @Test
-    void testLdayFromTraffic() {
-
-        SHPRead.importTable(connection, TestNoiseModelling.getResource("ROADS2.shp").getPath())
-
-        new Import_File().exec(connection,
-                ["pathFile" : TestNoiseModelling.getResource("buildings.shp").getPath(),
-                 "inputSRID": "2154",
-                 "tableName": "buildings"])
-
-        new Import_File().exec(connection,
-                ["pathFile" : TestNoiseModelling.getResource("receivers.shp").getPath(),
-                 "inputSRID": "2154",
-                 "tableName": "receivers"])
-
-
-        String res = new Noise_level_from_traffic().exec(connection,
-                ["tableBuilding" : "BUILDINGS",
-                 "tableRoads"    : "ROADS2",
-                 "tableReceivers": "RECEIVERS",
-                 "confMaxSrcDist" : 5000])
-
-        assertTrue(res.contains(NoiseMapDatabaseParameters.DEFAULT_RECEIVERS_LEVEL_TABLE_NAME))
-
-        def sql = new Sql(connection)
-
-        def periods = sql.rows("SELECT DISTINCT PERIOD FROM " + NoiseMapDatabaseParameters.DEFAULT_RECEIVERS_LEVEL_TABLE_NAME)
-        def periodValues = periods.collect {
-            it.PERIOD
-        }
-        assertTrue(periodValues.contains("D"))
-        assertTrue(periodValues.contains("E"))
-        assertTrue(periodValues.contains("N"))
-        assertTrue(periodValues.contains("DEN"))
-
-        def receiverCount = sql.firstRow("SELECT COUNT(*) CPT FROM RECEIVERS")["CPT"] as Integer
-
-        ["D", "E", "N", "DEN"].each { period ->
-            def periodCount = sql.firstRow("SELECT COUNT(*) CPT FROM " + NoiseMapDatabaseParameters.DEFAULT_RECEIVERS_LEVEL_TABLE_NAME + " WHERE PERIOD = ?", [period])["CPT"] as Integer
-            assertEquals(receiverCount, periodCount)
-        }
-
-        
-    }
-
-    @Test
-    void testLdenFromEmission() {
-
-        SHPRead.importTable(connection, TestNoiseModelling.getResource("ROADS2.shp").getPath())
-
-        new Road_Emission_from_Traffic().exec(connection,
-                ["tableRoads": "ROADS2"])
-
-        new Import_File().exec(connection,
-                ["pathFile" : TestNoiseModelling.getResource("buildings.shp").getPath(),
-                 "inputSRID": "2154",
-                 "tableName": "buildings"])
-
-        new Import_File().exec(connection,
-                ["pathFile" : TestNoiseModelling.getResource("receivers.shp").getPath(),
-                 "inputSRID": "2154",
-                 "tableName": "receivers"])
-
-
-        String res = new Noise_level_from_source().exec(connection,
-                ["tableBuilding"   : "BUILDINGS",
-                 "tableSources"   : "LW_ROADS",
-                 "tableReceivers": "RECEIVERS"])
-
-        assertTrue(res.contains(NoiseMapDatabaseParameters.DEFAULT_RECEIVERS_LEVEL_TABLE_NAME))
-    }
-
-    @Test
     void testLdenFromEmission1khz() {
 
         new Import_File().exec(connection,
